Tests/LibWeb/Text/expected/DOM/HTMLCollection-order.txt

HTMLCollection
anchor1
anchor2
anchor3
anchor1
anchor2
